Refactor suggestion / no-suggestion test into one
Jira ticket: CAMS-338
